refactor: #31681 address review suggestions from vincent-couturier-whoz
- Extract globToRegex to GlobUtils.kt in redirect package (reusable, single responsibility)
- Rename local variable eligible -> eligibleAgents in RedirectToolPlugin for clarity
- Replace if/else desc pattern with idiomatic Kotlin ?.also / run in RedirectTool.description
- Restore Dev.yaml GITHUB integration filters (add_issue_comment, issue_read/write, etc.)
  that were accidentally stripped in an earlier commit
