+++ agents/Dev.yaml
@@ -8,6 +8,13 @@ integrations:
   AI:
   MEMORY:
   GITHUB:
+    - add_issue_comment
+    - issue_read
+    - issue_write
+    - create_pull_request
+    - pull_request_read
+    - pull_request_review_write
+    - update_pull_request
   GIT:
   JIRA:
   DELEGATE:
